+++ docs/html/class_note.js
@@ -1,7 +1,7 @@
 var class_note =
 [
-    [ "Note", "class_note.html#acf27df9baa4638d2339fb7752f4e592f", null ],
-    [ "Note", "class_note.html#a7b651d3b2abd472c6afd769765c15809", null ],
+    [ "Note", "class_note.html#adbf69d0ddf6725df5b76511990d20e48", null ],
+    [ "Note", "class_note.html#a1fcf3688382b27c0ee15a262536deed8", null ],
     [ "setDurationTicks", "class_note.html#a469a9c7cc815626d43e901a71d5494bb", null ],
     [ "setIsNoteOn", "class_note.html#a7588b5da4cd36475994b7fd85620eb56", null ],
     [ "setOctave", "class_note.html#a5e9b1eef1992159f23c1744326c0a751", null ],