+++ docs/html/class_score.js
@@ -1,6 +1,6 @@
 var class_score =
 [
-    [ "Score", "class_score.html#adf2d37159d0ae9f633841ee7db3f3d53", null ],
+    [ "Score", "class_score.html#a014e053ce17389206a4e5a37cd391f7d", null ],
     [ "Score", "class_score.html#af957d12c7434f4f530aa173c178783f4", null ],
     [ "clear", "class_score.html#a0b19e00bea38679a5c627b7df3fe0cea", null ],
     [ "getChords", "class_score.html#aa4d6fd98b8ed74790a79e1c8fb095689", null ],