+++ docs/html/search/all_6.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['note_0',['Note',['../class_note.html',1,'Note'],['../class_note.html#acf27df9baa4638d2339fb7752f4e592f',1,'Note::Note(const std::string &amp;pitch, const size_t durationTicks=256, bool isNoteOn=true, bool inChord=false, int transposeDiatonic=0, int transposeChromatic=0, int divisionsPerQuarterNote=256)'],['../class_note.html#a7b651d3b2abd472c6afd769765c15809',1,'Note::Note(const size_t durationTicks=256, int divisionsPerQuarterNote=256)']]]
+  ['note_0',['Note',['../class_note.html',1,'Note'],['../class_note.html#adbf69d0ddf6725df5b76511990d20e48',1,'Note::Note(const std::string &amp;pitch, const size_t durationTicks=256, bool isNoteOn=true, bool inChord=false, int transposeDiatonic=0, int transposeChromatic=0)'],['../class_note.html#a1fcf3688382b27c0ee15a262536deed8',1,'Note::Note(const size_t durationTicks=256)']]]
 ];