+++ docs/html/search/functions_4.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['score_0',['Score',['../class_score.html#adf2d37159d0ae9f633841ee7db3f3d53',1,'Score::Score(const std::initializer_list&lt; std::string &gt; &amp;partsName, const int numMeasures=20, const int divisionsPerQuarterNote=256)'],['../class_score.html#af957d12c7434f4f530aa173c178783f4',1,'Score::Score(const std::string &amp;filePath)']]],
+  ['score_0',['Score',['../class_score.html#a014e053ce17389206a4e5a37cd391f7d',1,'Score::Score(const std::initializer_list&lt; std::string &gt; &amp;partsName, const int numMeasures=20)'],['../class_score.html#af957d12c7434f4f530aa173c178783f4',1,'Score::Score(const std::string &amp;filePath)']]],
   ['setdurationticks_1',['setDurationTicks',['../class_note.html#a469a9c7cc815626d43e901a71d5494bb',1,'Note']]],
   ['setisnoteon_2',['setIsNoteOn',['../class_note.html#a7588b5da4cd36475994b7fd85620eb56',1,'Note']]],
   ['setoctave_3',['setOctave',['../class_note.html#a5e9b1eef1992159f23c1744326c0a751',1,'Note']]],